% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_audio_features.R
\name{get_audio_features}
\alias{get_audio_features}
\title{Get audio features for specified track from API}
\usage{
get_audio_features(track_id, output = "tibble")
}
\arguments{
\item{track_id}{Valid track id from Spotify's API (or straight from the app if you want to do it that way instead).}

\item{output}{By default, results will be returned in a tibble. Any other value for this parameter will return the raw results.}
}
\value{
Results will be returned in a tibble by default with all tracks requested.
}
\description{
Get audio features for specified track from API
}
\examples{
get_my_top()
}
