% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_auth_header.R
\name{get_auth_header}
\alias{get_auth_header}
\title{Combine id/secret and return header.}
\usage{
get_auth_header(tkn, type = "user")
}
\arguments{
\item{tkn}{use saved token info to get header.}

\item{type}{either startup, app or user depending on the situation.}
}
\value{
object add_headers from httr with correct information.
}
\description{
Combine id/secret and return header.
}
\examples{
get_auth_header()
}
