% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_client_info.R
\name{get_client_info}
\alias{get_client_info}
\title{Pull client id and secret from environment vars.}
\usage{
get_client_info(type = "list")
}
\arguments{
\item{type}{return list by default. If set to 'auth64' return base64 encoded.}
}
\value{
list with client id and secret
}
\description{
Pull client id and secret from environment vars.
}
\examples{
get_client_info()
}
