% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_my_top.R
\name{get_my_top}
\alias{get_my_top}
\title{Combine id/secret and return header.}
\usage{
get_my_top(tkn, n = 20, type = "tracks", time_range = "short_term",
  output = "tbl")
}
\arguments{
\item{tkn}{use saved token info to get header.}

\item{n}{number of results to return.}

\item{type}{either 'tracks' or 'artists' depending on usage}

\item{time_range}{takes either 'short_term', 'medium_term' or 'long_term'. Set to 'short_term' by default.}

\item{output}{by default, will return a tibble object. Will allow 'raw' to return the raw endpoint call.}
}
\value{
tibble with relevant information
}
\description{
Combine id/secret and return header.
}
\examples{
get_my_top()
}
