% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_playlist_tracks.R
\name{get_playlist_tracks}
\alias{get_playlist_tracks}
\title{Get all tracks from playlist}
\usage{
get_playlist_tracks(tkn, playlist_id, output = "tbl")
}
\arguments{
\item{tkn}{User credentials. May need to spoof this if requesting a public playlist.}

\item{playlist_id}{Valid playlist ID from Spotify.}
}
\value{
Raw data returned from API
}
\description{
Get all tracks from playlist
}
\examples{
get_playlist_tracks()
}
