% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_registration_url.R
\name{get_registration_url}
\alias{get_registration_url}
\title{Combine id/secret and return header.}
\usage{
get_registration_url(scope = c("playlist-read-collaborative",
  "user-library-read", "user-follow-read", "user-read-email",
  "user-read-private", "user-top-read"))
}
\arguments{
\item{scope}{List describing the kind of access being requested.}
}
\value{
URL that will request access to a user's Spotify data.
}
\description{
Combine id/secret and return header.
}
\examples{
get_registration_url()
}
