% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_details.R
\name{get_user_details}
\alias{get_user_details}
\title{Retrieve User Details}
\usage{
get_user_details(tkn)
}
\arguments{
\item{tkn}{connection object including information on where to retrieve user's connection details}
}
\value{
User details from API
}
\description{
Given connection object, pull user details from API
}
\examples{
get_user_details()
}
