% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_users.R
\name{list_users}
\alias{list_users}
\title{List all users available to query}
\usage{
list_users()
}
\value{
user_list A list of all users. Can reference to pull token information.
}
\description{
Depends on having the SPOTIFYAPIDIR environment variable set.
}
\examples{
list_users()
}
