% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_user.R
\name{load_user}
\alias{load_user}
\title{Load user's authentication information into a variable for use.}
\usage{
load_user(username)
}
\arguments{
\item{username}{Just the users's username string.}
}
\value{
access_token Valid access token
}
\description{
Pull user information and retrieve an access_token using refresh endpoint.
}
\examples{
load_user()
}
