% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_user.R
\name{register_user}
\alias{register_user}
\title{Register User}
\usage{
register_user(redirect_uri)
}
\value{
User connection object to be used in further API requests.
}
\description{
Given full redirec_uri, extract code, register user and return user connection object
}
\examples{
register_user()
}
