% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotify_connect_refresh.R
\name{spotify_connect_refresh}
\alias{spotify_connect_refresh}
\title{Refresh access token given user info.}
\usage{
spotify_connect_refresh(tkn)
}
\arguments{
\item{tkn}{the user's authentication info as given by load_user}
}
\value{
tkn with new access token
}
\description{
Access tokens need to be refreshed every hour. This function provides a simple way to get a new access token given the users' refresh token.
}
\examples{
spotify_connect_refresh()
}
